pyproject.toml
[tool.poetry]
name = "shopper-sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["Joanna Khek <joannakhek@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
pandas = ">=1.3.5"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
kaggle = "^1.5.16"
pyyaml = "^6.0.1"
numpy = "^1.21.4"
torch = {version = "^2.1.1+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.16.0+cu118", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.1.1+cu118", source = "pytorch-gpu-src"}
torchtext = "^0.16.0"
torchinfo = "^1.8.0"
scikit-learn = "^1.3.2"
torchmetrics = "^1.2.1"
hydra-core = "^1.3.2"
spacy = "^3.7.2"
datasets = "^2.15.0"
transformers = {extras = ["torch"], version = "^4.36.2"}
nltk = "^3.8.1"
evaluate = "^0.4.1"
fastapi = {extras = ["all"], version = "^0.105.0"}
loguru = "^0.7.2"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.25.2"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"